[MOV] spreadheet: move global filter excel export
When exporting the global filters to excel, we inject a new sheet in
the exported excel data with the global filter values. But it was done
in the middle of the export, and this sheet does not exist in the model.
This caused the plugins that came after to try to use getters with this
mocked sheet. This worked by chance before, but now crashes.

This commit moves the injection of the global filter sheet to the
global filter feature plugin, that is at the very end of the export
process, so it doesn't impact other plugins.

Task: 6209932
