+++ addons/spreadsheet/static/src/global_filters/plugins/global_filters_core_view_plugin.js
@@ -23,16 +23,14 @@ import {
     getDateGlobalFilterValueFromDefault,
 } from "@spreadsheet/global_filters/helpers";
 import { OdooCoreViewPlugin } from "@spreadsheet/plugins";
-import { getItemId } from "../../helpers/model";
 import { serializeDate } from "@web/core/l10n/dates";
 import { getFilterCellValue, getFilterValueDomain } from "../helpers";
 const { DateTime } = luxon;
 
-const { UuidGenerator, createEmptyExcelSheet, createEmptySheet, toXC, toNumber } = helpers;
+const { toNumber } = helpers;
 
 export class GlobalFiltersCoreViewPlugin extends OdooCoreViewPlugin {
     static getters = /** @type {const} */ ([
-        "exportSheetWithActiveFilters",
         "getFilterDisplayValue",
         "getGlobalFilterByName",
         "getGlobalFilterDomain",
@@ -346,69 +344,5 @@ export class GlobalFiltersCoreViewPlugin extends OdooCoreViewPlugin {
             this.getters
         );
         return getDateDomain(from, to, field, type);
-    }
-
-    /**
-     * Adds all active filters (and their values) at the time of export in a dedicated sheet
-     *
-     * @param {Object} data
-     */
-    exportForExcel(data) {
-        if (this.getters.getGlobalFilters().length === 0) {
-            return;
-        }
-        this.exportSheetWithActiveFilters(data);
-        data.sheets[data.sheets.length - 1] = {
-            ...createEmptyExcelSheet(UuidGenerator.smallUuid(), _t("Active Filters")),
-            ...data.sheets.at(-1),
-        };
-    }
-
-    exportSheetWithActiveFilters(data) {
-        if (this.getters.getGlobalFilters().length === 0) {
-            return;
-        }
-
-        const cells = {
-            A1: "Filter",
-            B1: "Value",
-        };
-        const formats = {};
-        let numberOfCols = 2; // at least 2 cols (filter title and filter value)
-        let filterRowIndex = 1; // first row is the column titles
-        for (const filter of this.getters.getGlobalFilters()) {
-            cells[`A${filterRowIndex + 1}`] = filter.label;
-            const result = this.getFilterDisplayValue(filter.label);
-            for (const colIndex in result) {
-                numberOfCols = Math.max(numberOfCols, Number(colIndex) + 2);
-                for (const rowIndex in result[colIndex]) {
-                    const cell = result[colIndex][rowIndex];
-                    if (cell.value === undefined) {
-                        continue;
-                    }
-                    const xc = toXC(Number(colIndex) + 1, Number(rowIndex) + filterRowIndex);
-                    cells[xc] = cell.value.toString();
-                    if (cell.format) {
-                        const formatId = getItemId(cell.format, data.formats);
-                        formats[xc] = formatId;
-                    }
-                }
-            }
-            filterRowIndex += result[0].length;
-        }
-        const styleId = getItemId({ bold: true }, data.styles);
-
-        const sheet = {
-            ...createEmptySheet(UuidGenerator.smallUuid(), _t("Active Filters")),
-            cells,
-            formats,
-            styles: {
-                A1: styleId,
-                B1: styleId,
-            },
-            colNumber: numberOfCols,
-            rowNumber: filterRowIndex,
-        };
-        data.sheets.push(sheet);
     }
 }
